backend/requirements.txt
flask==3.1.*
flask-cors==5.*
pymysql==1.1.*
cryptography==44.*
pyjwt==2.10.*
bcrypt==4.3.*
python-dotenv==1.1.*
